api/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
distro==1.8.0
docstring-parser==0.15
Flask==2.3.3
frozenlist==1.4.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
idna==3.4
instructor==0.3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
multidict==6.0.4
openai==1.3.3
packaging==23.2
pydantic==2.5.1
pydantic_core==2.14.3
python-dotenv==1.0.0
requests==2.31.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.4
Werkzeug==2.3.7
yarl==1.9.2
